backend/requirements.txt
flask
gunicorn
firebase-admin
numpy
opencv-python-headless
pillow
requests
mtcnn
keras-facenet
tensorflow